Move PurchaseMetaExpiration component into its own tsx file
