+++ client/me/purchases/manage-purchase/purchase-meta.jsx
@@ -1,15 +1,11 @@
-import config from '@automattic/calypso-config';
 import {
 	isConciergeSession,
 	isDomainRegistration,
-	isDomainTransfer,
 	isJetpackPlan,
 	isJetpackProduct,
-	JETPACK_LEGACY_PLANS,
 } from '@automattic/calypso-products';
 import { Card } from '@automattic/components';
 import { localizeUrl } from '@automattic/i18n-utils';
-import classNames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
 import { times } from 'lodash';
 import PropTypes from 'prop-types';
@@ -17,28 +13,23 @@ import { useState, useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import ClipboardButton from 'calypso/components/forms/clipboard-button';
 import FormTextInput from 'calypso/components/forms/form-text-input';
-import { useLocalizedMoment } from 'calypso/components/localized-moment';
 import useUserLicenseBySubscriptionQuery from 'calypso/data/jetpack-licensing/use-user-license-by-subscription-query';
 import {
 	getName,
-	hasPaymentMethod,
 	isExpired,
 	isExpiring,
 	isIncludedWithPlan,
 	isOneTimePurchase,
-	isRechargeable,
 	isRenewing,
 	isSubscription,
-	isCloseToExpiration,
-	isRenewable,
 } from 'calypso/lib/purchases';
 import { CALYPSO_CONTACT, JETPACK_SUPPORT } from 'calypso/lib/url/support';
-import { getCurrentUser, getCurrentUserId } from 'calypso/state/current-user/selectors';
+import { getCurrentUser } from 'calypso/state/current-user/selectors';
 import { getByPurchaseId } from 'calypso/state/purchases/selectors';
 import { getSite, isRequestingSites } from 'calypso/state/sites/selectors';
 import { managePurchase } from '../paths';
 import { isJetpackTemporarySitePurchase } from '../utils';
-import AutoRenewToggle from './auto-renew-toggle';
+import PurchaseMetaExpiration from './purchase-meta-expiration';
 import PurchaseMetaIntroductoryOfferDetail from './purchase-meta-introductory-offer-detail';
 import PurchaseMetaOwner from './purchase-meta-owner';
 import PurchaseMetaPaymentDetails from './purchase-meta-payment-details';
@@ -106,84 +97,6 @@ PurchaseMeta.propTypes = {
 	getManagePurchaseUrlFor: PropTypes.func,
 	getChangePaymentMethodUrlFor: PropTypes.func,
 };
-
-function renderRenewsOrExpiresOnLabel( { purchase, translate } ) {
-	if ( isExpiring( purchase ) ) {
-		if ( isDomainRegistration( purchase ) ) {
-			return translate( 'Domain expires on' );
-		}
-
-		if ( isSubscription( purchase ) ) {
-			return translate( 'Subscription expires on' );
-		}
-
-		if ( isOneTimePurchase( purchase ) ) {
-			return translate( 'Expires on' );
-		}
-	}
-
-	if ( isExpired( purchase ) ) {
-		if ( isDomainRegistration( purchase ) ) {
-			return translate( 'Domain expired on' );
-		}
-
-		if ( isConciergeSession( purchase ) ) {
-			return translate( 'Session used on' );
-		}
-
-		if ( isSubscription( purchase ) ) {
-			return translate( 'Subscription expired on' );
-		}
-
-		if ( isOneTimePurchase( purchase ) ) {
-			return translate( 'Expired on' );
-		}
-	}
-
-	if ( isDomainRegistration( purchase ) ) {
-		return translate( 'Domain renews on' );
-	}
-
-	if ( isSubscription( purchase ) ) {
-		return translate( 'Subscription renews on' );
-	}
-
-	if ( isOneTimePurchase( purchase ) ) {
-		return translate( 'Renews on' );
-	}
-
-	return null;
-}
-
-function renderRenewsOrExpiresOn( {
-	moment,
-	purchase,
-	siteSlug,
-	translate,
-	getManagePurchaseUrlFor,
-} ) {
-	if ( isIncludedWithPlan( purchase ) ) {
-		const attachedPlanUrl = getManagePurchaseUrlFor( siteSlug, purchase.attachedToPurchaseId );
-
-		return (
-			<span>
-				<a href={ attachedPlanUrl }>{ translate( 'Renews with Plan' ) }</a>
-			</span>
-		);
-	}
-
-	if ( isExpiring( purchase ) || isExpired( purchase ) ) {
-		return moment( purchase.expiryDate ).format( 'LL' );
-	}
-
-	if ( isRenewing( purchase ) ) {
-		return moment( purchase.renewDate ).format( 'LL' );
-	}
-
-	if ( isOneTimePurchase( purchase ) ) {
-		return translate( 'Never Expires' );
-	}
-}
 
 function PurchaseMetaPlaceholder() {
 	return (
@@ -268,140 +181,6 @@ function RenewErrorMessage( { purchase, translate, site } ) {
 				}
 			) }
 		</div>
-	);
-}
-
-function PurchaseMetaExpiration( {
-	purchase,
-	site,
-	siteSlug,
-	getChangePaymentMethodUrlFor,
-	getManagePurchaseUrlFor,
-} ) {
-	const translate = useTranslate();
-	const moment = useLocalizedMoment();
-	const isProductOwner = purchase?.userId === useSelector( getCurrentUserId );
-	const isJetpackPurchase = isJetpackPlan( purchase ) || isJetpackProduct( purchase );
-	const isAutorenewalEnabled = purchase?.isAutoRenewEnabled ?? false;
-	const isJetpackPurchaseUsingPrimaryCancellationFlow =
-		isJetpackPurchase && config.isEnabled( 'jetpack/cancel-through-main-flow' );
-	const hideAutoRenew =
-		purchase && JETPACK_LEGACY_PLANS.includes( purchase.productSlug ) && ! isRenewable( purchase );
-
-	if ( ! purchase || isDomainTransfer( purchase ) || purchase?.isInAppPurchase ) {
-		return null;
-	}
-
-	if ( isRenewable( purchase ) && ! isExpired( purchase ) ) {
-		const dateSpan = <span className="manage-purchase__detail-date-span" />;
-		// If a jetpack site has been disconnected, the "site" prop will be null here.
-		const shouldRenderToggle = site && isProductOwner;
-		const autoRenewToggle = shouldRenderToggle ? (
-			<AutoRenewToggle
-				planName={ site.plan?.product_name_short ?? '' }
-				siteDomain={ site.domain }
-				siteSlug={ site.slug }
-				purchase={ purchase }
-				toggleSource="manage-purchase"
-				showLink={ true }
-				getChangePaymentMethodUrlFor={ getChangePaymentMethodUrlFor }
-			/>
-		) : (
-			<span />
-		);
-		const subsRenewText = isAutorenewalEnabled
-			? translate( 'Auto-renew is {{autoRenewToggle}}ON{{/autoRenewToggle}}', {
-					components: {
-						autoRenewToggle,
-					},
-			  } )
-			: translate( 'Auto-renew is {{autoRenewToggle}}OFF{{/autoRenewToggle}}', {
-					components: {
-						autoRenewToggle,
-					},
-			  } );
-
-		const subsReEnableText = translate(
-			'{{autoRenewToggle}}Re-activate subscription{{/autoRenewToggle}}',
-			{
-				components: {
-					autoRenewToggle,
-				},
-			}
-		);
-
-		let subsBillingText;
-		if (
-			isAutorenewalEnabled &&
-			! hideAutoRenew &&
-			hasPaymentMethod( purchase ) &&
-			isRechargeable( purchase )
-		) {
-			subsBillingText = translate( 'You will be billed on {{dateSpan}}%(renewDate)s{{/dateSpan}}', {
-				args: {
-					renewDate: purchase.renewDate && moment( purchase.renewDate ).format( 'LL' ),
-				},
-				components: {
-					dateSpan,
-				},
-			} );
-		} else {
-			subsBillingText = translate( 'Expires on {{dateSpan}}%(expireDate)s{{/dateSpan}}', {
-				args: {
-					expireDate: moment( purchase.expiryDate ).format( 'LL' ),
-				},
-				components: {
-					dateSpan,
-				},
-			} );
-		}
-
-		return (
-			<li className="manage-purchase__meta-expiration">
-				<em className="manage-purchase__detail-label">{ translate( 'Subscription Renewal' ) }</em>
-				{ ! hideAutoRenew && ! isJetpackPurchaseUsingPrimaryCancellationFlow && (
-					<div className="manage-purchase__auto-renew">
-						<span className="manage-purchase__detail manage-purchase__auto-renew-text">
-							{ subsRenewText }
-						</span>
-					</div>
-				) }
-				<span
-					className={ classNames( 'manage-purchase__detail', {
-						'is-expiring': isCloseToExpiration( purchase ),
-					} ) }
-				>
-					{ subsBillingText }
-				</span>
-				{ ! isAutorenewalEnabled &&
-					! hideAutoRenew &&
-					shouldRenderToggle &&
-					isJetpackPurchaseUsingPrimaryCancellationFlow && (
-						<div className="manage-purchase__auto-renew">
-							<span className="manage-purchase__detail manage-purchase__auto-renew-text">
-								{ subsReEnableText }
-							</span>
-						</div>
-					) }
-			</li>
-		);
-	}
-
-	return (
-		<li>
-			<em className="manage-purchase__detail-label">
-				{ renderRenewsOrExpiresOnLabel( { purchase, translate } ) }
-			</em>
-			<span className="manage-purchase__detail">
-				{ renderRenewsOrExpiresOn( {
-					moment,
-					purchase,
-					siteSlug,
-					translate,
-					getManagePurchaseUrlFor,
-				} ) }
-			</span>
-		</li>
 	);
 }
